test: merge options to option
